move config methods to separate file
